[feature] add global flag for disabling the events framework
this will be used to allow the unittests to disable the events
framework. this way, emit() will become a passthrough.

note that, until now, the basetest class is making use of the threaded
version of the client, which launches a zmq tornado-based ioloop. this
is wrong, and will have to be addressed in a future commit. we'll have
to make use of the global EVENTS_ENABLED flag in the txclient version
when those changes are made.

Related: #7259
Relases: 0.4.2