[pkg] remove dependency on dirspec
This commit removes the dep introduced in 5e12233 by just importing some tiny
bit of dirspec code.

The previous change was introduced because:

  * pyxdg did not account for Mac OS specifics, i.e. using ~/Library/
    directory structure instead of .config (see:
    https://leap.se/code/issues/3574).

  * dirspec does the correct thing for xdg on Mac OS.

  * u1db depends on dirspec anyway.

The problem is that dirspec is not maintained and published on pypi, what
forces us to download it from an URL and add exceptions to be able to pip
install it.

As we are removing dependence on u1db on other modules, we can also remove it
here. To workaround the Mac OS problem, we just add some code from dirspec to
ensure we get the correct directory on Mac OS.